refactor(base): organize preview components
