+++ packages/@sanity/base/src/components/previews/general/DefaultPreview.tsx
@@ -3,7 +3,7 @@ import {Box, Flex, Stack, Text, Skeleton, TextSkeleton} from '@sanity/ui'
 import styled from 'styled-components'
 import {getDevicePixelRatio} from 'use-device-pixel-ratio'
 import classNames from 'classnames'
-import {PreviewProps} from './types'
+import {PreviewProps} from '../types'
 
 const Root = styled(Flex)`
   height: 35px;
