+++ packages/@sanity/base/src/components/previews/general/DetailPreview.styled.tsx
