+++ packages/@sanity/base/src/components/previews/general/MediaPreview.styled.tsx
