+++ packages/@sanity/base/src/components/previews/general/MediaPreview.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import {Box, Text} from '@sanity/ui'
 import {getDevicePixelRatio} from 'use-device-pixel-ratio'
-import {ProgressCircle} from '../progress'
-import {MediaDimensions, PreviewProps} from './types'
-import {MediaWrapper, MediaString, Root, ProgressWrapper} from './mediaPreview.styled'
+import {ProgressCircle} from '../../progress'
+import {MediaDimensions, PreviewProps} from '../types'
+import {MediaWrapper, MediaString, Root, ProgressWrapper} from './MediaPreview.styled'
 
 const DEFAULT_MEDIA_DIMENSIONS: MediaDimensions = {
   width: 160,
