+++ packages/@sanity/base/src/components/previews/portableText/BlockImagePreview.styled.tsx
