+++ packages/@sanity/base/src/components/previews/template/CreateDocumentPreview.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import {Box, Flex, Stack, Text} from '@sanity/ui'
 import styled from 'styled-components'
-import {IntentButton} from '../IntentButton'
-import {MediaDimensions} from './types'
+import {IntentButton} from '../../IntentButton'
+import {MediaDimensions} from '../types'
 
 interface CreateDocumentPreviewProps {
   title?: React.ReactNode | React.FunctionComponent<unknown>
